pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coupa-mcp"
version = "0.1.0"
description = "MCP server wrapping the Coupa Procurement REST API"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "httpx>=0.28.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
coupa-mcp = "coupa_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/coupa_mcp"]
